[fix] 데미지 뷰어 함수 매개변수의 자료형 변경

+++ Client/Assets/Scripts/Managers/Core/UIManager.cs
@@ -175,7 +175,7 @@ public class UIManager
         _itemTooltip?.HideTooltip();
     }
 
-    public void ShowDamageText(float damage, Vector3 worldPosition)
+    public void ShowDamageText(int damage, Vector3 worldPosition)
     {
         UI_DamageViewer damageViewer = Managers.UI.MakeWorldSpaceUI<UI_DamageViewer>();
         damageViewer.ShowDamage(damage, worldPosition);
